Added pyvalexec symlink, added --version.
pyval_exec.py is symlinked to pyvalexec,
so you can just run ./pyvalexec or even
pyvalexec (if it's in /usr/bin or likewise).

added -v,--version flags to command-line args.
if it's going to be a stand-alone command it needs
to be more well-behaved and not look like it was
just glued together to run from the command line.