Added ability to load code from a file.
You can now pass a filename to pyval_exec,
it will try to load the contents and run the
code. The name __main__ is not set when running,
so scripts don't behave EXACTLY like they would
in an interpreter. So far every statement I've
thrown at it gets executed.

Updated version to 1.0.4