Alternate config,Default channel,Forward privmsgs.
An alternate config file may be used by passing the
filename using command-line options.

The new default channel is '##<bot nick>', which
complies with FreeNode rules for channels.

Privmsgs are now forwarded to admins.

New commands were added to allow messaging all admins,
or all channels that the bot is in.

Other code-organiztion was done. Mainly to the bot's
channel-topic stuff.